new data runner added, 2 new folders
final proposition for the structure also encompasses two new folders for environment and data_collector, these were initially there in the last commit to 151 as well however due to them being called env and data they were ignored when pushed onto the branch on git, a git ignored is what caused that. However after I saw them out of the folders I thought it might also be pointless to put one file in one folder each? I dont know, up to you if this is how it'll look like or not